Update translate.cpp
I have added the function implementations for Translating and Formatting. All of the functions utilize string pointer to simplify the main program, but I can easily change it to return a string if we want. An interesting thing to note is: because translate has no return call, any word will not be translated if it does not have a translation. This makes it work well for proper nouns such as names. If you put "Nick is smart" it cannot translate "Nick" and will put that into the translation and write "nick es inteligente". Just a cool thing to note.